Extract samples shared elements
