+++ samples/WebApi.Mvc/Startup.cs
@@ -16,6 +16,9 @@ namespace WebApi.Mvc
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
+            //services.AddExceptionMapper(builder => builder
+            //    .MapCommonMvcExceptions()
+            //);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
