+++ samples/WebApi.Shared/GoneException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace WebApi.Shared
+{
+    public class GoneException : Exception
+    {
+    }
+}
