+++ samples/WebApi.Shared/ImATeapotException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace WebApi.Shared
+{
+    public class ImATeapotException : Exception
+    {
+
+    }
+}
