+++ samples/WebApi.Shared/MyForbiddenException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace WebApi.Shared
+{
+    public class MyForbiddenException : Exception
+    {
+
+    }
+}
