+++ samples/WebApi.Shared/MyForbiddenExceptionHandler.cs
@@ -1,0 +1,10 @@
+﻿using ForEvolve.ExceptionMapper;
+using Microsoft.AspNetCore.Http;
+
+namespace WebApi.Shared
+{
+    public class MyForbiddenExceptionHandler : ExceptionHandler<MyForbiddenException>
+    {
+        public override int StatusCode => StatusCodes.Status403Forbidden;
+    }
+}
