+++ samples/WebApi.Shared/MyNotFoundException.cs
@@ -1,0 +1,8 @@
+﻿using ForEvolve.ExceptionMapper;
+
+namespace WebApi.Shared
+{
+    public class MyNotFoundException : NotFoundException
+    {
+    }
+}
