+++ samples/WebApi.Shared/MyUnauthorizedException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace WebApi.Shared
+{
+    public class MyUnauthorizedException : Exception
+    {
+    }
+}
